Fix test dependencies.
