+++ admin-webapp/src/main/java/org/opentestsystem/rdw/admin/ApplicationConfiguration.java
@@ -1,11 +1,32 @@
 package org.opentestsystem.rdw.admin;
 
+import org.opentestsystem.rdw.admin.repository.DataSourceConfiguration;
 import org.opentestsystem.rdw.admin.web.NotFoundIndexTemplate;
+import org.opentestsystem.rdw.common.status.StatusConfiguration;
+import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
+import org.opentestsystem.rdw.reporting.common.web.WebConfiguration;
+import org.opentestsystem.rdw.reporting.common.web.security.SecurityConfiguration;
+import org.opentestsystem.rdw.security.repository.JdbcOrganizationRepository;
+import org.opentestsystem.rdw.utils.ResourceLoaderConfiguration;
+import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
+import org.springframework.boot.actuate.autoconfigure.EndpointAutoConfiguration;
 import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 @Configuration
+@Import({
+        ResourceLoaderConfiguration.class,
+        EndpointAutoConfiguration.class,
+        StatusConfiguration.class,
+        YamlPropertiesConfigurator.class,
+        DataSourceConfiguration.class,
+        SecurityConfiguration.class,
+        JdbcOrganizationRepository.class,
+        WebConfiguration.class,
+        SecurityParameterProvider.class
+})
 public class ApplicationConfiguration {
 
     @Bean
