+++ admin-webapp/src/test/java/org/opentestsystem/rdw/admin/web/StudentGroupBatchControllerIT.java
@@ -6,10 +6,12 @@ import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.admin.model.ImportStatus;
 import org.opentestsystem.rdw.admin.model.StudentGroupBatch;
 import org.opentestsystem.rdw.admin.service.StudentGroupBatchService;
+import org.opentestsystem.rdw.reporting.common.web.ExceptionHandlerControllerAdvice;
 import org.opentestsystem.rdw.reporting.common.web.security.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Import;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -29,6 +31,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @RunWith(SpringRunner.class)
 @WebMvcTest(StudentGroupBatchController.class)
+@Import(ExceptionHandlerControllerAdvice.class)
 public class StudentGroupBatchControllerIT {
 
     @Autowired
