+++ common-web/src/main/java/org/opentestsystem/rdw/reporting/common/web/security/SecurityConfiguration.java
@@ -1,6 +1,7 @@
 package org.opentestsystem.rdw.reporting.common.web.security;
 
 import org.opentestsystem.rdw.reporting.common.web.security.session.SecuredRedisConfiguration;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
@@ -12,5 +13,10 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
         SecuredRedisConfiguration.class
 })
 public class SecurityConfiguration {
+
+    @Bean
+    public Object mySuperBean() {
+        return new Object();
+    }
 
 }
