+++ report-processor/src/main/java/org/opentestsystem/rdw/reporting/processor/ApplicationConfiguration.java
@@ -1,0 +1,44 @@
+package org.opentestsystem.rdw.reporting.processor;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.opentestsystem.rdw.common.status.StatusConfiguration;
+import org.opentestsystem.rdw.reporting.common.i18n.TranslationConfiguration;
+import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
+import org.opentestsystem.rdw.reporting.common.repository.DataSourceConfiguration;
+import org.opentestsystem.rdw.reporting.processor.service.WkhtmltopdfWebServiceClient;
+import org.opentestsystem.rdw.utils.ResourceLoaderConfiguration;
+import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.actuate.autoconfigure.EndpointAutoConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.http.client.ClientHttpRequestFactory;
+import org.springframework.http.client.SimpleClientHttpRequestFactory;
+
+@Configuration
+@Import({
+        EndpointAutoConfiguration.class,
+        ResourceLoaderConfiguration.class,
+        StatusConfiguration.class,
+        YamlPropertiesConfigurator.class,
+        DataSourceConfiguration.class,
+        TranslationConfiguration.class,
+        SecurityParameterProvider.class
+})
+public class ApplicationConfiguration {
+
+    @Bean
+    public ClientHttpRequestFactory requestFactory() {
+        return new SimpleClientHttpRequestFactory();
+    }
+
+    @Bean
+    public WkhtmltopdfWebServiceClient wkhtmlToPdfWebServiceClient(
+            @Value("${app.wkhtmltopdf.url}") final String url,
+            final ClientHttpRequestFactory requestFactory,
+            final ObjectMapper objectMapper) {
+        return new WkhtmltopdfWebServiceClient(requestFactory, url, objectMapper);
+    }
+
+}
