+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/Application.java
@@ -1,23 +1,10 @@
 package org.opentestsystem.rdw.reporting;
 
-import org.opentestsystem.rdw.common.status.StatusConfiguration;
-import org.opentestsystem.rdw.reporting.common.i18n.TranslationConfiguration;
-import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
-import org.opentestsystem.rdw.reporting.common.repository.DataSourceConfiguration;
-import org.opentestsystem.rdw.reporting.common.web.WebConfiguration;
-import org.opentestsystem.rdw.reporting.common.web.security.SecurityConfiguration;
-import org.opentestsystem.rdw.reporting.iris.IrisConfiguration;
-import org.opentestsystem.rdw.reporting.report.ReportConfiguration;
-import org.opentestsystem.rdw.security.repository.JdbcOrganizationRepository;
-import org.opentestsystem.rdw.utils.ResourceLoaderConfiguration;
-import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.actuate.autoconfigure.EndpointAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cloud.aws.autoconfigure.cache.ElastiCacheAutoConfiguration;
 import org.springframework.cloud.aws.autoconfigure.context.ContextStackAutoConfiguration;
-import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.PropertySource;
 
 @SpringBootApplication(exclude = {
@@ -26,20 +13,6 @@ import org.springframework.context.annotation.PropertySource;
 })
 @EnableCaching
 @PropertySource(value = "classpath:build-info.properties", ignoreResourceNotFound = true)
-@Import({
-        ResourceLoaderConfiguration.class,
-        EndpointAutoConfiguration.class,
-        StatusConfiguration.class,
-        YamlPropertiesConfigurator.class,
-        DataSourceConfiguration.class,
-        TranslationConfiguration.class,
-        ReportConfiguration.class,
-        SecurityConfiguration.class,
-        WebConfiguration.class,
-        JdbcOrganizationRepository.class,
-        SecurityParameterProvider.class,
-        IrisConfiguration.class
-})
 public class Application {
 
     public static void main(final String[] args) {
