+++ webapp/src/main/java/org/opentestsystem/rdw/reporting/ApplicationConfiguration.java
@@ -1,10 +1,37 @@
 package org.opentestsystem.rdw.reporting;
 
+import org.opentestsystem.rdw.common.status.StatusConfiguration;
+import org.opentestsystem.rdw.reporting.common.i18n.TranslationConfiguration;
+import org.opentestsystem.rdw.reporting.common.jdbc.SecurityParameterProvider;
+import org.opentestsystem.rdw.reporting.common.repository.DataSourceConfiguration;
+import org.opentestsystem.rdw.reporting.common.web.WebConfiguration;
+import org.opentestsystem.rdw.reporting.common.web.security.SecurityConfiguration;
+import org.opentestsystem.rdw.reporting.iris.IrisConfiguration;
+import org.opentestsystem.rdw.reporting.report.ReportConfiguration;
+import org.opentestsystem.rdw.security.repository.JdbcOrganizationRepository;
+import org.opentestsystem.rdw.utils.ResourceLoaderConfiguration;
+import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
+import org.springframework.boot.actuate.autoconfigure.EndpointAutoConfiguration;
 import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 @Configuration
 @EnableZuulProxy
+@Import({
+        ResourceLoaderConfiguration.class,
+        EndpointAutoConfiguration.class,
+        StatusConfiguration.class,
+        YamlPropertiesConfigurator.class,
+        DataSourceConfiguration.class,
+        TranslationConfiguration.class,
+        ReportConfiguration.class,
+        SecurityConfiguration.class,
+        WebConfiguration.class,
+        JdbcOrganizationRepository.class,
+        SecurityParameterProvider.class,
+        IrisConfiguration.class
+})
 public class ApplicationConfiguration {
 
 }
